requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
configobj==5.0.8
configparser==6.0.1
dataclasses-json==0.6.4
distro==1.9.0
etelemetry==0.3.1
exceptiongroup==1.2.0
faiss-cpu==1.8.0
filelock==3.13.3
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.42
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.6
isodate==0.6.1
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
langchain==0.1.13
langchain-community==0.0.29
langchain-core==0.1.35
langchain-text-splitters==0.0.1
langsmith==0.1.34
looseversion==1.3.0
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.2.1
nibabel==5.2.1
nipype==1.8.6
numpy==1.26.4
openai==1.14.3
orjson==3.9.15
packaging==23.2
pandas==2.2.1
pathlib==1.0.1
pillow==10.2.0
protobuf==4.25.3
prov==2.0.0
pyarrow==15.0.2
pydantic==2.6.4
pydantic_core==2.16.3
pydeck==0.8.1b0
pydot==2.0.0
Pygments==2.17.2
PyMuPDF==1.24.0
PyMuPDFb==1.24.0
pyparsing==3.1.2
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyxnat==1.6.2
PyYAML==6.0.1
rdflib==7.0.0
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
scipy==1.12.0
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.29
streamlit==1.32.2
tenacity==8.2.3
tiktoken==0.6.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
traits==6.3.2
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
yarl==1.9.4
python-docx==1.1.2
